unique pw removed, removed supers, removed some default files, moved pages to pages folder
